code cleanup 1: Seperated routes into their own file. Added coroutines for authentication
