+++ src/data/todos/MySqlToDoRepository.kt
@@ -3,39 +3,76 @@ package com.dev.james.data.todos
 import com.dev.james.database.DatabaseManager
 import com.dev.james.entities.ToDo
 import com.dev.james.entities.ToDoDraft
+import kotlinx.coroutines.CoroutineScope
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.async
+import kotlinx.coroutines.coroutineScope
 
 class MySqlToDoRepository : ToDoRepository {
-    private val database = DatabaseManager()
 
-    override fun getAllTodos(userId : String): List<ToDo> {
-        return database.getAllToDosByUserId(userId)
-            .map { ToDo(it.id, it.title, it.done , it.user_id) }
+    private val database = DatabaseManager()
+    private val scope = CoroutineScope(Dispatchers.IO)
+
+    override suspend fun getAllTodos(userId : String): List<ToDo> {
+
+        val dataResponse = scope.async {
+            database.getAllToDosByUserId(userId)
+                .map { ToDo(it.id, it.title, it.done , it.user_id) }
+        }
+
+        return dataResponse.await()
     }
 
-    override fun getToDo(id: Int , userId: String): ToDo? {
-        return database.getTodo(id , userId)?.let {
+    override suspend fun getToDo(id: Int, userId: String): ToDo? {
+
+            val result = scope.async {
+                database.getTodo(id , userId)?.let {
+                    ToDo(
+                        it.id,
+                        it.title,
+                        it.done ,
+                        it.user_id
+                    )
+                }
+            }
+
+        return result.await()
+
+        /*return database.getTodo(id , userId)?.let {
             ToDo(
                 it.id,
                 it.title,
                 it.done ,
                 it.user_id
             )
-        }
+        }*/
+
     }
 
-    override fun addToDo(draft: ToDoDraft): ToDo {
-        return database
-            .addTodo(draft)
+    override suspend fun addToDo(draft: ToDoDraft): ToDo {
+
+            val todoDeferred = scope.async {
+                database
+                    .addTodo(draft)
+            }
+
+        return todoDeferred.await()
     }
 
-    override fun removeToDo(id: Int , userId: String): Boolean {
-        return database.removeToDo(id , userId)
+    override suspend fun removeToDo(id: Int, userId: String): Boolean {
+            val isRemovedDeferred = scope.async {
+                database.removeToDo(id , userId)
+            }
+        return isRemovedDeferred.await()
     }
 
-    override fun updateToDo(id: Int, draft: ToDoDraft): Boolean {
-        return database.updateTodo(
-            id,
-            draft
-        )
+    override suspend fun updateToDo(id: Int, draft: ToDoDraft): Boolean {
+            val isUpdated = scope.async {
+                database.updateTodo(
+                    id,
+                    draft
+                )
+            }
+        return isUpdated.await()
     }
 }
